stgi-ecommerce/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.7
django-filter==2.4.0
django-cors-headers==3.8.0
djangorestframework==3.12.4
pycodestyle==2.7.0
pytz==2021.1
sqlparse==0.4.2
toml==0.10.2
psycopg2-binary==2.9.9